first working version of migration

+++ applications/dtw/dtw.h
@@ -9,10 +9,10 @@
 #define p4          4
 #define recognizer  5
 
-#define MATX_SIZE           11       // tamanho da matriz
-#define TOTAL_TASKS         4      // deve ser PAR para dividir igualmente o numero de padroes por task
-#define NUM_PATTERNS        400 // 40
-#define PATTERN_PER_TASK    NUM_PATTERNS / TOTAL_TASKS
+#define MATX_SIZE           11      // tamanho da matriz
+#define TOTAL_TASKS         4       // deve ser PAR para dividir igualmente o numero de padroes por task
+#define NUM_PATTERNS        400     // 
+#define PATTERN_PER_TASK    (int)NUM_PATTERNS / TOTAL_TASKS
 #define MESSAGE_MAX_SIZE    128
 
 static unsigned int P[TOTAL_TASKS] = {p1, p2, p3, p4};
